init go1.15
