+++ compiler/natives/src/internal/poll/go115_utils.go
@@ -1,0 +1,12 @@
+// +build js
+// +build go1.15
+
+package poll
+
+import (
+	"errors"
+)
+
+var (
+	ErrTimeout = errors.New("timeout")
+)
