+++ compiler/natives/src/runtime/go115_utils.go
@@ -1,0 +1,8 @@
+// +build js
+// +build go1.15
+
+package runtime
+
+func throw(s string) {
+	panic(s)
+}
