+++ compiler/natives/src/runtime/utils.go
@@ -1,0 +1,12 @@
+// +build js
+// +build !go1.15
+
+package runtime
+
+func efaceOf(ep *interface{}) *eface {
+	panic("efaceOf: not supported")
+}
+
+func throw(s string) {
+	panic(errorString(s))
+}
