+++ compiler/natives/src/time/timer.go
@@ -1,0 +1,17 @@
+// +build js
+// +build !go1.15
+
+package time
+
+import "github.com/gopherjs/gopherjs/js"
+
+type runtimeTimer struct {
+	i       int32
+	when    int64
+	period  int64
+	f       func(interface{}, uintptr)
+	arg     interface{}
+	timeout *js.Object
+	active  bool
+	req     uintptr
+}
